Make `html` a parameter to the main function
The function is more stand-alone and testable if `html` is passed in as a parameter, rather than being read off the page. This mirrors the way that `html` is returned at the end of the function, rather than it being directly assigned to `document.body.innerHTML`.

It is not strictly necessary to create a new variable `newHTML` – reassigning to the parameter `html` does not change the value that was passed into the function. But since JavaScript’s behavior in this case is not obvious, I made the code more explicit so readers wouldn’t worry about the possibility of that bug.